+++ src/rr.jl
@@ -142,5 +142,5 @@ function predict(object::Rr, X; lb::Union{Nothing, Float64, AbstractVector{Float
         pred[i] = z.int .+ X * z.B
     end 
     if le_lb == 1 ; pred = pred[1] ; end
-    (pred = pred,)
+    (pred = pred, lb)
 end